+++ lib/util/logger.ts
@@ -1,0 +1,5 @@
+export function devWarn(msg: string) {
+  if(process.env.NODE_ENV === "development") {
+    console.warn(msg)
+  }
+}