+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+# [v0.2.0]
+
+## Breaking
+
+- Rename fields name
+  - `apiKey` => `client_id`
+  - `apiSecretKey` => `client_secret`
+
+## Changes
+
+- Adding api client error handle, api error will perform `console.warn` in development mode(NODE_ENV=development)
+
 # [v0.1.1]
 
-- add video fields title
+- Add video fields title
 
 # [v0.1.0]
 
-- video api CRUD
+- Video api CRUD